requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
discord-protos==0.0.2
discord.py-self==2.0.0
frozenlist==1.4.1
gTTS==2.5.1
idna==3.7
multidict==6.0.5
protobuf==5.27.2
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
